requirements.txt
asgiref==3.4.1
backports.entry-points-selectable==1.1.0
distlib==0.3.2
dj-database-url==0.5.0
Django==3.2.7
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
filelock==3.0.12
gunicorn==20.1.0
platformdirs==2.3.0
psycopg2==2.9.1
PyJWT==2.1.0
python-dotenv==0.19.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.2
virtualenv==20.7.2
virtualenvwrapper-win==1.2.6
whitenoise==5.3.0
